searching for subestations done

+++ apps/subestaciones/models.py
@@ -3,6 +3,6 @@ from django.db import models
 # Create your models here.
 class Subestacion(models.Model):
     nombre = models.CharField(max_length=50, null=False, unique=True)
-    latitud = models.FloatField()
-    longitud = models.FloatField()
+    latitud = models.DecimalField(max_digits=18, decimal_places=15)
+    longitud = models.DecimalField(max_digits=18, decimal_places=15)
     activo = models.BooleanField(default=True)
